Moving views and routes to app_server
